requirements.txt
asgiref==3.4.1
Django==3.2.12
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==8.4.0
pkg-resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
